+++ packages/services/f-analytics/src/mixins/_tests/analytics.methods.test.js
@@ -5,28 +5,25 @@ import * as utils from '../../utils/helpers';
 import {
     createStore,
     $route,
-    $i18n
+    $i18n,
+    $cookies
 } from '../../tests/helpers/setup';
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
 
 describe('Analytics', () => {
-    afterEach(() => {
-        jest.clearAllMocks();
-    });
-
     describe('methods ::', () => {
-        // 1 == locale, 2 == branding, 3 == country, 4 == currency, 5 == language
-        const cases = [
-            ['en-GB', 'justeat', 'uk', 'gbp', 'en'],
-            ['en-IE', 'justeat', 'ie', 'eur', 'en'],
-            ['it-IT', 'justeat', 'it', 'eur', 'it'],
-            ['es-ES', 'justeat', 'es', 'eur', 'es'],
-            ['en-AU', 'menulog', 'au', 'aud', 'en'],
-            ['en-NZ', 'menulog', 'nz', 'nzd', 'en']
-        ];
         describe('preparePlatformData ::', () => {
+            // 1 == locale, 2 == branding, 3 == country, 4 == currency, 5 == language
+            const cases = [
+                ['en-GB', 'justeat', 'uk', 'gbp', 'en'],
+                ['en-IE', 'justeat', 'ie', 'eur', 'en'],
+                ['it-IT', 'justeat', 'it', 'eur', 'it'],
+                ['es-ES', 'justeat', 'es', 'eur', 'es'],
+                ['en-AU', 'menulog', 'au', 'aud', 'en'],
+                ['en-NZ', 'menulog', 'nz', 'nzd', 'en']
+            ];
             let component;
             let storeUpdatePlatformDataSpy;
 
@@ -34,7 +31,6 @@ describe('Analytics', () => {
                 component = {
                     render () {},
                     mixins: [analyticsMixin],
-                    store: createStore(),
                     computed: {
                         isServer () {
                             return false;
@@ -89,6 +85,74 @@ describe('Analytics', () => {
                     expect(storeUpdatePlatformDataSpy).toHaveBeenCalledWith(expected);
                 }
             );
+        });
+
+        describe('prepareUserData ::', () => {
+            const userIdFromCookie = 'fjdhskgshjgk';
+
+            const mockAuthToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.'
+            + 'eyJlbWFpbCI6ImpvZS5ibG9nZ3NAanVzdGVhdHRha2Vhd2F5LmNvbS'
+            + 'IsImNyZWF0ZWRfZGF0ZSI6IjIwMjEtMDItMDhUMTA6Mjc6NDkuMTkz'
+            + 'MDAwMFoiLCJuYW1lIjoiSm9lIEJsb2dncyIsImdsb2JhbF91c2VyX2lkI'
+            + 'joiVTdOUkFsV0FnNXpPZHNkUmdmN25rVHlvaTkwWEVvPSIsImdpdmVuX25h'
+            + 'bWUiOiJKb2UiLCJmYW1pbHlfbmFtZSI6IkJsb2dncyIsImlhdCI6MTYxNTQ2OTUxNn0.VapH6uHnn4lHIkvN_mS9A9IVVWL0YPNE39gDDD-l7SU';
+
+            const userDataWithMockAuthToken = {
+                'a-UserId': userIdFromCookie,
+                authType: 'Login',
+                email: '1a9a31f72fbb57efd148bbfe06c169b97f6868200b422a5ae7fed7e3f853002a',
+                globalUserId: 'U7NRAlWAg5zOdsdRgf7nkTyoi90XEo=',
+                signinType: 'Email',
+                signupDate: '2021-02-08T10:27:49.1930000Z'
+            };
+
+            const userDataWithoutMockAuthToken = {
+                'a-UserId': userIdFromCookie,
+                authType: undefined,
+                email: undefined,
+                globalUserId: undefined,
+                signinType: undefined,
+                signupDate: undefined
+            };
+
+            const component = {
+                render () {},
+                mixins: [analyticsMixin]
+            };
+            let wrapper;
+
+            component.mixins[0].created = jest.fn(() => true);
+            const storeUpdateUserDataSpy = jest.spyOn(component.mixins[0].methods, 'updateUserData');
+
+            beforeEach(() => {
+                wrapper = shallowMount(component, {
+                    mixins: [analyticsMixin],
+                    localVue,
+                    store: createStore(),
+                    mocks: {
+                        $i18n,
+                        $cookies
+                    }
+                });
+
+                wrapper.vm.$cookies.get.mockReturnValue(userIdFromCookie);
+            });
+
+            it('should call `updateUserData` action only with userId if authToken has not been passed', () => {
+                // Act
+                wrapper.vm.prepareUserData();
+
+                // Assert
+                expect(storeUpdateUserDataSpy).toHaveBeenCalledWith(userDataWithoutMockAuthToken);
+            });
+
+            it('should call `updateUserData` action with useauthToken data if authToken has been passed', () => {
+                // Act
+                wrapper.vm.prepareUserData(mockAuthToken);
+
+                // Assert
+                expect(storeUpdateUserDataSpy).toHaveBeenCalledWith(userDataWithMockAuthToken);
+            });
         });
 
         describe('preparePageData ::', () => {
@@ -182,6 +246,27 @@ describe('Analytics', () => {
                 // Assert
                 expect(storeUpdatePageDataSpy).toHaveBeenCalledWith(expected);
             });
+
+            it('should add conversationId and requestId values when passed in param object', () => {
+                // Arrange
+                const expected = {
+                    ...basePageDataObject,
+                    conversationId: '460cc3a8-83f7-4e80-bb46-c8a69967f249',
+                    requestId: '6cbe6509-9122-4e66-a90a-cc483c34282e'
+                };
+                jest.spyOn(utils, 'getDisplaySize').mockImplementation(() => expected.display);
+                jest.spyOn(utils, 'getOrientation').mockImplementation(() => expected.orientation);
+                jest.spyOn(utils, 'mapRouteToGroup').mockImplementation(() => expected.group);
+                jest.spyOn(utils, 'mapRouteToFeature').mockImplementation(() => expected.name);
+
+                // Act
+                wrapper.vm.preparePageData({ conversationId: expected.conversationId, requestId: expected.requestId });
+
+                // Assert
+                expect(storeUpdatePageDataSpy).toHaveBeenCalledWith(expected);
+            });
         });
     });
 });
+
+